+++ internal/configs/testdata/providers_count/root.tf
@@ -1,0 +1,11 @@
+terraform {
+  required_providers {
+    local = {
+      source  = "hashicorp/local"
+    }
+  }
+}
+
+provider "foo-test" {
+  count = 3
+}